docs: add autostart instructions to README
Add instructions on how to configure automatic authentication on systemd-based Linux distros in the README file. The instructions include details on modifying the configuration files and setting permissions for the program to access the config file.

docs(systemd/user): add user systemd unit files for autostart

Add user systemd unit files for automatic authentication on systemd-based Linux distros. The unit files are located in the `docs/systemd/user` folder. Users can modify the path in the configuration files and copy them to the appropriate location.